feat: edit custom fields
