+++ apps/server/src/services/rundown-service/rundownCache.ts
@@ -1,4 +1,5 @@
 import {
+  CustomField,
   CustomFieldLabel,
   CustomFields,
   isOntimeDelay,
@@ -17,10 +18,11 @@ type EventID = string;
 type NormalisedRundown = Record<EventID, OntimeRundownEntry>;
 
 let persistedRundown: OntimeRundown = [];
+const persistedCustomFields: CustomFields = {};
 
-/** Utility function gets rundown from DataProvider */
+/** Utility function gets to expose data */
 export const getPersistedRundown = (): OntimeRundown => persistedRundown;
-export const getCustomFields = (): CustomFields => testCustomProperties;
+export const getCustomFields = (): CustomFields => persistedCustomFields;
 
 let rundown: NormalisedRundown = {};
 let order: EventID[] = [];
@@ -30,7 +32,17 @@ let totalDelay = 0;
 
 let links: Record<EventID, EventID> = {};
 
-const assignedCustomProperties: Record<CustomFieldLabel, EventID[]> = {};
+/**
+ * Object that contains renamings to custom fields
+ * Used to rename the custom fields in the events
+ * @example
+ * {
+ *  oldLabel: newLabel
+ *  lighting: lx
+ * }
+ */
+const customFieldChangelog = {};
+const assignedCustomFields: Record<CustomFieldLabel, EventID[]> = {};
 
 export async function init(initialRundown: OntimeRundown) {
   persistedRundown = structuredClone(initialRundown);
@@ -38,26 +50,13 @@ export async function init(initialRundown: OntimeRundown) {
   await DataProvider.setRundown(persistedRundown);
 }
 
-const testCustomProperties: CustomFields = {
-  lighting: {
-    colour: 'red',
-    label: 'lighting',
-    type: 'string',
-  },
-  sound: {
-    colour: 'amber',
-    label: 'sound',
-    type: 'string',
-  },
-};
-
 /**
  * Utility initialises cache
  * @param rundown
  */
 export function generate(
   initialRundown: OntimeRundown = persistedRundown,
-  customProperties: CustomFields = testCustomProperties,
+  customProperties: CustomFields = persistedCustomFields,
 ) {
   // we decided to re-write this dataset for every change
   // instead of maintaining logic to update it
@@ -107,10 +106,10 @@ export function generate(
             delete updatedEvent.custom[property];
             return;
           }
-          if (!Array.isArray(assignedCustomProperties[property])) {
-            assignedCustomProperties[property] = [];
+          if (!Array.isArray(assignedCustomFields[property])) {
+            assignedCustomFields[property] = [];
           }
-          assignedCustomProperties[property].push(updatedEvent.id);
+          assignedCustomFields[property].push(updatedEvent.id);
         }
         // update the persisted event
         initialRundown[i] = updatedEvent;
@@ -138,7 +137,7 @@ export function generate(
 
   isStale = false;
   totalDelay = accumulatedDelay;
-  return { rundown, order, links, totalDelay, assignedCustomProperties };
+  return { rundown, order, links, totalDelay, assignedCustomProperties: assignedCustomFields };
 }
 
 /** Returns an ID guaranteed to be unique */
@@ -278,6 +277,7 @@ export function edit({ persistedRundown, eventId, patch }: EditArgs): Required<M
 
   const eventInMemory = persistedRundown[indexAt];
   const newEvent = makeEvent(eventInMemory, patch);
+  console.log('got', patch, 'will make', newEvent);
 
   const newRundown = [...persistedRundown];
   newRundown[indexAt] = newEvent;
@@ -353,3 +353,73 @@ export function swap({ persistedRundown, fromId, toId }: SwapArgs): MutatingRetu
 
   return { newRundown };
 }
+
+/**
+ * Sanitises and creates a custom field in the database
+ * @param field
+ * @returns
+ */
+export const createCustomField = async (field: CustomField) => {
+  const { label, type, colour } = field;
+
+  // check if label already exists
+  const alreadyExists = Object.hasOwn(persistedCustomFields, label);
+
+  if (alreadyExists) {
+    throw new Error('Label already exists');
+  }
+
+  // update object and persist
+  persistedCustomFields[label] = { label, type, colour };
+
+  setImmediate(() => {
+    DataProvider.setCustomFields(persistedCustomFields);
+  });
+
+  return persistedCustomFields;
+};
+
+/**
+ * Edits an existing custom field in the database
+ * @param label
+ * @param newField
+ * @returns
+ */
+export const editCustomField = async (label: string, newField: Partial<CustomField>) => {
+  if (!(label in persistedCustomFields)) {
+    throw new Error('Could not find label');
+  }
+
+  const existingField = persistedCustomFields[label];
+  if (existingField.type !== newField.type) {
+    throw new Error('Change of field type is not allowed');
+  }
+
+  if (existingField.label !== newField.label) {
+    customFieldChangelog[label] = newField.label;
+  }
+
+  persistedCustomFields[label] = { ...existingField, ...newField };
+
+  setImmediate(() => {
+    DataProvider.setCustomFields(persistedCustomFields);
+  });
+
+  return persistedCustomFields;
+};
+
+/**
+ * Deletes a custom field from the database
+ * @param label
+ */
+export const removeCustomField = async (label: string) => {
+  if (label in persistedCustomFields) {
+    delete persistedCustomFields[label];
+  }
+
+  setImmediate(() => {
+    DataProvider.setCustomFields(persistedCustomFields);
+  });
+
+  return persistedCustomFields;
+};
